+++ src/helpers/logger/blank.js
