+++ src/helpers/logger/index.js
@@ -2,6 +2,6 @@ module.exports.colorlog = require('./colorlog')
 module.exports.loader = require('./loader')
 module.exports.error = require('./error')
 module.exports.done = require('./done')
-module.exports.noop = require('./noop')
+module.exports.blank = require('./blank')
 module.exports.exit = require('./exit')
 module.exports.finish = require('./finish')
